requirements.txt
google-cloud-pubsub==2.15.0
google-cloud-logging==3.5.0